requirements.txt
yagmail[all]
RPIO
picamera